Clean up companion object APIs
