+++ src/main/kotlin/com/slack/auto/value/kotlin/AutoValueKotlinExtension.kt
@@ -19,6 +19,8 @@ package com.slack.auto.value.kotlin
 
 import com.google.auto.service.AutoService
 import com.google.auto.value.extension.AutoValueExtension
+import com.google.auto.value.extension.AutoValueExtension.BuilderContext
+import com.slack.auto.value.kotlin.AvkBuilder.BuilderProperty
 import com.squareup.kotlinpoet.AnnotationSpec
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.CodeBlock
@@ -26,6 +28,7 @@ import com.squareup.kotlinpoet.DelicateKotlinPoetApi
 import com.squareup.kotlinpoet.FunSpec
 import com.squareup.kotlinpoet.KModifier
 import com.squareup.kotlinpoet.PropertySpec
+import com.squareup.kotlinpoet.TypeName
 import com.squareup.kotlinpoet.asClassName
 import com.squareup.kotlinpoet.asTypeVariableName
 import com.squareup.kotlinpoet.joinToCode
@@ -429,4 +432,65 @@ public class AutoValueKotlinExtension : AutoValueExtension() {
 
   private fun TypeMirror.isClassOfType(types: Types, other: TypeMirror?) =
     types.isAssignable(this, other)
+}
+
+private fun AvkBuilder.Companion.from(
+  builderContext: BuilderContext,
+  propertyTypes: Map<String, TypeName>,
+  parseDocs: Element.() -> String?
+): AvkBuilder {
+  // Setters
+  val props = builderContext.setters().entries.map { (prop, setters) ->
+    val type = propertyTypes.getValue(prop)
+    BuilderProperty(
+      prop,
+      type,
+      setters.mapTo(LinkedHashSet()) {
+        FunSpec.copyOf(it)
+          .withDocsFrom(it, parseDocs)
+          .build()
+      }
+    )
+  }
+
+  val builderMethods = builderContext.setters().values.flatten()
+    .plus(builderContext.autoBuildMethod())
+    .toMutableSet()
+
+  if (builderContext.buildMethod().isPresent) {
+    builderMethods += builderContext.buildMethod().get()
+  }
+
+  // TODO propertyBuilders
+
+  val remainingMethods =
+    ElementFilter.methodsIn(builderContext.builderType().enclosedElements)
+      .asSequence()
+      .filterNot { it in builderMethods }
+      .filterNot { it == builderContext.autoBuildMethod() }
+      .map { "${it.modifiers.joinToString(" ")} ${it.returnType} ${it.simpleName}(...)" }
+      .toList()
+
+  return AvkBuilder(
+    name = builderContext.builderType().simpleName.toString(),
+    doc = builderContext.builderType().parseDocs(),
+    visibility = if (PUBLIC in builderContext.builderType().modifiers) {
+      KModifier.PUBLIC
+    } else {
+      KModifier.INTERNAL
+    },
+    builderProps = props,
+    buildFun = builderContext.buildMethod()
+      .map {
+        FunSpec.copyOf(it)
+          .withDocsFrom(it, parseDocs)
+          .build()
+      }
+      .orElse(null),
+    autoBuildFun = FunSpec.copyOf(builderContext.autoBuildMethod())
+      .withDocsFrom(builderContext.autoBuildMethod(), parseDocs)
+      .build(),
+    remainingMethods = remainingMethods,
+    classAnnotations = builderContext.builderType().classAnnotations()
+  )
 }
